Support Scala Native
