+++ .gitignore
@@ -10,3 +10,4 @@ project/.sbtserver
 tags
 nohup.out
 out
+lowered.hnir
