+++ appveyor.yml
@@ -5,9 +5,9 @@ os: Windows Server 2012 R2
 environment:
   matrix:
   - JAVA_HOME: C:\Program Files\Java\jdk1.8.0
-    TEST_TASKS: __[2.11.12].test __[2.12.7].test
+    TEST_TASKS: __.jvm.test
   - JAVA_HOME: C:\Program Files\Java\jdk9
-    TEST_TASKS: __[2.11.12].test __[2.12.7].test
+    TEST_TASKS: __.jvm.test
 
 init:
   # Appveyor has so,e problems with symlinks..
@@ -15,7 +15,7 @@ init:
   - git config --global core.symlinks true
 
 install:
-  - SET MILL_URL=https://github.com/lihaoyi/mill/releases/download/0.3.3/0.3.3
+  - SET MILL_URL=https://github.com/lihaoyi/mill/releases/download/0.6.1/0.6.1
 
 build_script:
   - SET "PATH=%JAVA_HOME%\bin;%PATH%"
