+++ os/src-jvm/PlatformSpecific.scala
@@ -1,0 +1,7 @@
+package os
+
+private[os] trait PlatformSpecific {
+  def resource(implicit resRoot: ResourceRoot = Thread.currentThread().getContextClassLoader) ={
+    os.ResourcePath.resource(resRoot)
+  }
+}
