+++ os/src-native/PlatformSpecific.scala
@@ -1,0 +1,3 @@
+package os
+
+private[os] trait PlatformSpecific {}
